fix(beam/sdks/go): adding todos and moving code to correct places
