--- sdks/go/pkg/beam/xlangx.go
@@ -1,60 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one or more
-// contributor license agreements.  See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
-// The ASF licenses this file to You under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
-// the License.  You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package beam
-
-import (
-	"github.com/apache/beam/sdks/go/pkg/beam/core/graph"
-)
-
-func PCollectionToNode(p PCollection) *graph.Node {
-	if !p.IsValid() {
-		panic("tried converting invalid PCollection")
-	}
-	return p.n
-}
-
-func NodeToPCollection(n *graph.Node) PCollection {
-	if n == nil {
-		panic("tried converting invalid Node")
-	}
-	c := PCollection{n}
-	c.SetCoder(NewCoder(c.Type()))
-	return c
-}
-
-func mapPCollectionToNode(pMap map[string]PCollection) map[string]*graph.Node {
-	if pMap == nil {
-		return nil
-	}
-
-	nMap := make(map[string]*graph.Node)
-	for k, p := range pMap {
-		nMap[k] = PCollectionToNode(p)
-	}
-	return nMap
-}
-
-func mapNodeToPCollection(nMap map[string]*graph.Node) map[string]PCollection {
-	if nMap == nil {
-		return nil
-	}
-
-	pMap := make(map[string]PCollection)
-	for k, n := range nMap {
-		pMap[k] = NodeToPCollection(n)
-	}
-	return pMap
-}
